+++ documentation/changelog.md
@@ -14,6 +14,7 @@ description: List of all the changes and current versions.
 + Smooth Login Transitions
 + Smoother Leaderboards
 + Hide Scrollbar on View Instance Creation
++ Hide Screen Until Spawn
 ```
 
 ## 1.0.6 - January 4-5, 2021